minor adjustments

+++ srcs/ft_str/ft_strjoin.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strjoin.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: victode- <victode-@student.42lisboa.com    +#+  +:+       +#+        */
+/*   By: victode- <victode-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/26 15:25:40 by victode-          #+#    #+#             */
-/*   Updated: 2025/08/26 15:54:34 by victode-         ###   ########.fr       */
+/*   Updated: 2025/10/08 20:30:48 by victode-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 	char	*join;
 
 	if (!s1 || !s2)
